Move multiple tests to pytest
